+++ internal/pkg/util/fs/overlay/overlay_set_linux.go
@@ -11,9 +11,12 @@ package overlay
 
 import (
 	"fmt"
+	"os"
+	"os/exec"
 	"strings"
 	"syscall"
 
+	"github.com/apptainer/apptainer/internal/pkg/util/bin"
 	"github.com/apptainer/apptainer/pkg/sylog"
 	"github.com/samber/lo"
 )
@@ -43,6 +46,15 @@ type Set struct {
 // directory.
 func (s Set) Mount(rootFsDir string) error {
 	// Perform identity mounts for this Set
+	dups := lo.FindDuplicatesBy(s.ReadonlyOverlays, func(item *Item) string {
+		return item.SourcePath
+	})
+	if len(dups) > 0 {
+		return fmt.Errorf("duplicate overlays detected: %v", lo.Map(dups, func(item *Item, _ int) string {
+			return item.SourcePath
+		}))
+	}
+
 	if err := s.performIndividualMounts(); err != nil {
 		return err
 	}
@@ -53,7 +65,18 @@ func (s Set) Mount(rootFsDir string) error {
 
 // UnmountOverlay ummounts a Set from a specified rootfs directory.
 func (s Set) Unmount(rootFsDir string) error {
-	if err := DetachMount(rootFsDir); err != nil {
+	unprivOls, err := UnprivOverlaysSupported()
+	if err != nil {
+		return fmt.Errorf("while checking for unprivileged overlay support in kernel: %w", err)
+	}
+
+	if unprivOls {
+		err = DetachMount(rootFsDir)
+	} else {
+		err = UnmountWithFuse(rootFsDir)
+	}
+
+	if err != nil {
 		return err
 	}
 
@@ -84,9 +107,36 @@ func (s Set) performIndividualMounts() error {
 func (s Set) performFinalMount(rootFsDir string) error {
 	// Try to perform actual mount
 	options := s.options(rootFsDir)
-	sylog.Debugf("Mounting overlay with rootFsDir %q, options: %q", rootFsDir, options)
-	if err := syscall.Mount("overlay", rootFsDir, "overlay", syscall.MS_NODEV, options); err != nil {
-		return fmt.Errorf("failed to mount %s: %s", rootFsDir, err)
+	unprivOls, err := UnprivOverlaysSupported()
+	if err != nil {
+		return fmt.Errorf("while checking for unprivileged overlay support in kernel: %w", err)
+	}
+
+	if unprivOls {
+		sylog.Debugf("Mounting overlay (via syscall) with rootFsDir %q, options: %q", rootFsDir, options)
+		if err := syscall.Mount("overlay", rootFsDir, "overlay", syscall.MS_NODEV, options); err != nil {
+			return fmt.Errorf("failed to mount %s: %w", rootFsDir, err)
+		}
+	} else {
+		fuseOlFsCmd, err := bin.FindBin("fuse-overlayfs")
+		if err != nil {
+			return fmt.Errorf("kernel does not support unprivileged overlays, and fuse-overlayfs not available: %w", err)
+		}
+
+		// Even though fusermount is not needed for this step, we shouldn't perform
+		// the mount unless we have the necessary tools to eventually unmount it
+		_, err = bin.FindBin("fusermount")
+		if err != nil {
+			return fmt.Errorf("kernel does not support unprivileged overlays, and using fuse-overlayfs fallback requires fusermount to be installed: %w", err)
+		}
+
+		sylog.Debugf("Mounting overlay (via fuse-overlayfs) with rootFsDir %q, options: %q", rootFsDir, options)
+		execCmd := exec.Command(fuseOlFsCmd, "-o", options, rootFsDir)
+		execCmd.Stderr = os.Stderr
+		_, err = execCmd.Output()
+		if err != nil {
+			return fmt.Errorf("failed to mount %s: %w", rootFsDir, err)
+		}
 	}
 
 	return nil